requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.114.0
greenlet==3.0.3
h11==0.14.0
idna==3.8
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2==2.9.9
pydantic==2.9.1
pydantic_core==2.23.3
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
python-dotenv==1.0.1